requirements.txt
alembic==1.18.4
amqp==5.3.1
annotated-types==0.7.0
billiard==4.2.4
blinker==1.9.0
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-Mailman==1.1.1
flask-marshmallow==1.4.0
Flask-Migrate==4.1.0
Flask-Principal==0.4.0
Flask-Security-Too==5.7.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai @ git+https://github.com/google/generative-ai-python.git@7a7cc5474ddaa0255a4410e05361028a24400abd
googleapis-common-protos==1.73.1
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.71.2
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.2
libpass==1.9.3
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.2.2
marshmallow-sqlalchemy==1.4.2
packaging==26.0
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
redis==6.4.0
requests==2.33.0
setuptools==82.0.1
six==1.17.0
SQLAlchemy==2.0.48
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
Werkzeug==3.1.7
WTForms==3.2.1
